+++ CHANGES.rst
@@ -13,6 +13,7 @@ next
 - Provide syslog logging ability. #56.
 - Add path verification component for integration testing. #61.
 - Include vanguards_parallel.sh for high load services. #62.
+- Set NumDirectoryGuards to 2, to help ensure only 2 orconns. #63
 
 0.3.1
 -----